requirements.txt
aiohttp==3.8.1
disnake==2.5.2
python-dotenv==0.20.0
requests
